% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs.R
\name{apply_cohort_labels}
\alias{apply_cohort_labels}
\title{Add labels to cohort variable groups}
\usage{
apply_cohort_labels(df, demographic_vector)
}
\arguments{
\item{df}{the output of compare_cohort_def including all variables computed}

\item{demographic_vector}{a vector of demographic categories provided in
demographic_mappings; ensures all demographics are captured under the appropriate label}
}
\value{
returns the same input dataframe with an additional fact_group column
that labels each of the computed variables for easier grouping
}
\description{
Add labels to cohort variable groups
}
