% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs_pcornet.R
\name{compare_cohort_def_pcnt}
\alias{compare_cohort_def_pcnt}
\title{Reformat cohort definition tables}
\usage{
compare_cohort_def_pcnt(
  base_cohort,
  alt_cohorts,
  multi_or_single_site,
  person_tbl = cdm_tbl("demographic"),
  visit_tbl = cdm_tbl("encounter"),
  provider_tbl = NULL,
  care_site_tbl = NULL,
  demographic_mappings = sensitivityselectioncriteria::ssc_pcornet_demographics,
  specialty_concepts = NULL,
  outcome_concepts = NULL,
  domain_defs = sensitivityselectioncriteria::ssc_domain_file,
  domain_select = sensitivityselectioncriteria::ssc_domain_file \%>\% distinct(domain)
    \%>\% pull()
)
}
\arguments{
\item{base_cohort}{table with the base cohort & inclusion criteria; must include
site, person_id, start_date, end_date}

\item{alt_cohorts}{list of each table with an alternative inclusion criteria definition
applied to the base_cohort}

\item{multi_or_single_site}{direction to determine what kind of check to run
string that is either \code{multi} or \code{single}}

\item{person_tbl}{the CDM table that contains patient information}

\item{visit_tbl}{the CDM table that contains visit/encounter information}

\item{provider_tbl}{the CDM table with provider & provider specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{care_site_tbl}{the CDM table with care site / facility & care site / facility specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{demographic_mappings}{table defining how demographic elements should be defined
if NULL, the default demographic mappings for the CDM will be used
(\code{ssc_pcornet_demographics})
otherwise, the user provided table will be used}

\item{specialty_concepts}{a concept set with provider specialty concepts of interest
to be used to identify specialty visits}

\item{outcome_concepts}{a concept set with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{variable} | \code{domain}

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    where domain matches a domain listed in the `domain_defs` table
}\if{html}{\out{</div>}}}

\item{domain_defs}{a table with domain definitions with the following columns:
domain, domain_tbl, concept_field, date_field, filter_logic}

\item{domain_select}{a vector of domain names that should be used in the computation
of facts per domain per year of follow up

\if{html}{\out{<div class="sourceCode">}}\preformatted{                 this vector does NOT need to include domains used in the computation for
                 outcomes, as those will be accessed from the table directly
}\if{html}{\out{</div>}}}
}
\value{
one combined dataframe with one row per patient in the original cohort
with flags to show which cohort definitions apply to the patient.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    flag columns will include `base_cohort` and `alt_cohort_#` columns with numbers
    appended that correspond to the tables position in the `def_tbls` list

    `base_cohort` should equal 1 for all patients
}\if{html}{\out{</div>}}
}
\description{
Reformat cohort definition tables
}
\keyword{internal}
