% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs.R
\name{compare_cohort_smd}
\alias{compare_cohort_smd}
\title{Compute Standardized Mean Difference between cohort definitions}
\usage{
compare_cohort_smd(cohort_def_output, demographic_vector)
}
\arguments{
\item{cohort_def_output}{the output of compare_cohort_def}

\item{demographic_vector}{a vector of demographic categories provided in
demographic_mappings; ensures all demographics are captured under the appropriate label}
}
\value{
a summarized dataframe where the standardized mean difference between each
alternate cohort definition and the base definition is computed for each of
the computed variables
}
\description{
Compute Standardized Mean Difference between cohort definitions
}
\keyword{internal}
