% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs.R
\name{compute_domains_ssc}
\alias{compute_domains_ssc}
\title{Compute domain facts per person year}
\usage{
compute_domains_ssc(cohort, site_col, grouped_list, domain_tbl)
}
\arguments{
\item{cohort}{the combined cohort with patients from the base and alternate
cohort definitions, with a cohort_id label to differentiate}

\item{site_col}{the column in cohort with the site name(s)}

\item{grouped_list}{a list of variables for grouping}

\item{domain_tbl}{a table with domain definitions that contains the following columns:
domain, domain_tbl, concept_field, date_field, filter_logic}
}
\value{
a dataframe with fact counts of the user-provided domains per year of follow-up
within the cohort period for each of the patients in the cohort
}
\description{
Compute domain facts per person year
}
