% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs.R
\name{find_outcomes_ssc}
\alias{find_outcomes_ssc}
\title{Find presence of outcomes in cohort definitions}
\usage{
find_outcomes_ssc(cohort, site_col, domain_tbl, outcome_concepts)
}
\arguments{
\item{cohort}{the combined cohort with patients from the base and alternate
cohort definitions, with a cohort_id label to differentiate}

\item{site_col}{the column in the cohort_tbl with the site name(s)}

\item{domain_tbl}{a table with domain definitions with the following columns:
domain, domain_tbl, concept_field, date_field, filter_logic}

\item{outcome_concepts}{a concept set with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{domain}

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    where domain matches a domain listed in the domain_tbl
}\if{html}{\out{</div>}}}
}
\value{
a dataframe with indicators for each patient identifying whether or not they
have the outcome(s) of interest
}
\description{
Find presence of outcomes in cohort definitions
}
\keyword{internal}
