% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs.R
\name{find_specialty_visits}
\alias{find_specialty_visits}
\title{Compute specialty visits per person year}
\usage{
find_specialty_visits(
  cohort,
  site_col,
  specialty_concepts,
  grouped_list,
  provider_tbl = cdm_tbl("provider"),
  visit_tbl = cdm_tbl("visit_occurrence")
)
}
\arguments{
\item{cohort}{the combined cohort with patients from the base and alternate
cohort definitions, with a cohort_id label to differentiate}

\item{site_col}{the column in the cohort_tbl with the site name(s)}

\item{specialty_concepts}{a concept set with provider specialty concepts of interest
to be used to identify specialty visits}

\item{grouped_list}{a list of variables for grouping}

\item{provider_tbl}{the CDM table with provider & provider specialty information}

\item{visit_tbl}{The CDM table with visit/encounter information}
}
\value{
a dataframe with the number of specialist visits per year of follow up
for each of the patients in the cohort
}
\description{
Compute specialty visits per person year
}
