% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssc_domain_file}
\alias{ssc_domain_file}
\title{SSC Sample Domain File}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 6 columns.
}
\usage{
ssc_domain_file
}
\description{
A sample version of the file structure expected for the \code{domain_tbl}
parameter in the \code{ssc_process} function. The user should recreate
this file and include their own domain definitions.
}
\details{
#' @format ## ssc_domain_file
\describe{
A data frame with 2 rows and 5 columns
\item{domain}{An arbitrary string to label the domain of interest}
\item{domain_tbl}{The name of the CDM table associated with the domain of interest}
\item{concept_field}{The name of the column in the domain table that contains the concepts of interest listed in the concept_set file.}
\item{date_field}{The name of the column in the domain table that contains dates to be used for time-based filtering.}
\item{vocabulary_field}{(PCORnet only) The name of the column in the domain table where the vocabulary type is stored}
\item{filter_logic}{(optional) a string to be parsed as logic to filter the domain_tbl as needed to best represent the domain}
}
}
\keyword{datasets}
