% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_output_subfuncs.R
\name{ssc_ms_anom_nt}
\alias{ssc_ms_anom_nt}
\title{\emph{Multi-Site, Anomaly Detection, No Time}}
\usage{
ssc_ms_anom_nt(process_output)
}
\arguments{
\item{process_output}{the output of ssc_process}
}
\value{
a table displaying the mean, median, and iqr standardized mean difference
values for each site & alternative cohort definition vs baseline
a dot plot showing the standardized mean difference value vs baseline for
each site & alternative cohort definition
}
\description{
\emph{Multi-Site, Anomaly Detection, No Time}
}
