% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_output_subfuncs.R
\name{ssc_ms_exp_cs}
\alias{ssc_ms_exp_cs}
\title{\emph{Multi-Site, Exploratory, Cross-Sectional}}
\usage{
ssc_ms_exp_cs(process_output, alt_cohort_filter)
}
\arguments{
\item{process_output}{the output from ssc_process}

\item{alt_cohort_filter}{an vector indicating which alternate cohorts should be
displayed; only 2 are allowed at once to maintain visibility
in the graph}
}
\value{
two bump charts - one with the continuous facts (that are represented by medians) and
another with the categorical facts (that are represented by proportions)
if two alternate cohort definitions are provided, the base cohort is placed in the middle
of the plot. Otherwise, the base is on the left.
}
\description{
\emph{Multi-Site, Exploratory, Cross-Sectional}
}
