% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_output_subfuncs.R
\name{ssc_ms_exp_nt}
\alias{ssc_ms_exp_nt}
\title{\emph{Multi-Site, Exploratory, No Time}}
\usage{
ssc_ms_exp_nt(process_output)
}
\arguments{
\item{process_output}{the output from ssc_process}
}
\value{
a heat map showing the median facts per person year for each site in each cohort definition
a bar graph showing the proportion of patients with demographic and outcome characteristics
for each cohort definition & site
}
\description{
\emph{Multi-Site, Exploratory, No Time}
}
