% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssc_omop_demographics}
\alias{ssc_omop_demographics}
\title{SSC Demographic Mappings -- OMOP}
\format{
\subsection{ssc_omop_demographics}{

A data frame with 3 columns
\describe{
\item{demographic}{a string used to identify the demographic category}
\item{concept_field}{the name of the field in the CDM person table containing the appropriate demographic information}
\item{field_values}{the values in concept_field that represent the demographic category}
}
}
}
\usage{
ssc_omop_demographics
}
\description{
A sample version of the file structure expected for the \code{demographic_mappings}
parameter in the \code{ssc_process} function. The user should recreate this
file and include their own clinical concepts.
}
\details{
If the parameter is left NULL, this file or its PCORnet compliment will be used
in place of custom mappings.
}
\keyword{datasets}
