% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process.R
\name{ssc_process}
\alias{ssc_process}
\title{Sensitivity to Selection Criteria}
\usage{
ssc_process(
  base_cohort,
  alt_cohorts,
  omop_or_pcornet,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  person_tbl = cdm_tbl("person"),
  visit_tbl = cdm_tbl("visit_occurrence"),
  provider_tbl = NULL,
  care_site_tbl = NULL,
  demographic_mappings = NULL,
  specialty_concepts = NULL,
  outcome_concepts = NULL,
  domain_tbl = sensitivityselectioncriteria::ssc_domain_file,
  domain_select = sensitivityselectioncriteria::ssc_domain_file \%>\% distinct(domain)
    \%>\% pull()
)
}
\arguments{
\item{base_cohort}{\emph{tabular input} || \strong{required}

A table representing all patients who meet the base cohort definition, used
as the "gold standard" for comparative analyses. This table should have at least:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{alt_cohorts}{\emph{tablular input or list of tabular inputs} || \strong{required}

A table or list of tables (can be named or unnamed) representing patients
meeting alternative cohort definitions, which will be assessed against the
base cohort. These tables should have the same structure as the \code{base_cohort.}

If an unnamed list is provided, numbers will be assigned to each cohort
definition for labeling purposes. For named lists, the names will be used to
label the alternate cohorts.}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data
\itemize{
\item \code{omop}: run the \code{\link[=ssc_process_omop]{ssc_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=ssc_process_pcornet]{ssc_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{person_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('person')}

The CDM \code{person} or \code{demographic} table that contains basic demographic information
(sex, race, etc.) about the cohort members}

\item{visit_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('visit_occurrence')}

The CDM \code{visit_occurrence}, \code{visit_detail}, or \code{encounter} table that contains
information about the patients' healthcare encounters}

\item{provider_tbl}{\emph{tabular input} || defaults to \code{NULL}

The CDM table with provider & provider specialty information. This table is
only used if a \code{specialty_concepts} concept set is provided.}

\item{care_site_tbl}{\emph{tabular input} || defaults to \code{NULL}

The CDM table with care site/facility & related specialty information. This table is
only used if a \code{specialty_concepts} concept set is provided.}

\item{demographic_mappings}{\emph{tabular input} || defaults to \code{NULL}

A table defining how demographic elements should be defined. If left as NULL,
the default demographic mappings for the CDM will be used, which can be viewed
in \code{ssc_omop_demographics} and \code{ssc_pcornet_demographics}. If a different table is
provided by the user, those definition will be used instead. This table should
minimally contain:
\itemize{
\item \code{demographic} | \emph{character} | the label for the demographic category (ex: female, asian_race)
\item \code{concept_field} | \emph{character} | the field in the CDM where this information is stored
\item \code{field_values} | \emph{character} | the concept or concepts that are used to define the demographic category
}}

\item{specialty_concepts}{\emph{tabular input} || defaults to \code{NULL}

A concept set with concepts representing the specialties of interest to be
identified for the cohort. This table should minimally contain either
the \code{concept_id} field (OMOP) or the \code{concept_code} field (PCORnet)}

\item{outcome_concepts}{\emph{tabular input} || defaults to \code{NULL}

A concept set used to define any relevant outcome variables to be assessed
in each cohort. This input should contain one of following:
\itemize{
\item \code{concept_id} | \emph{integer} | the concept_id of interest (required for OMOP)
\item \code{concept_code} | \emph{character} | the code of interest (required for PCORnet)
}

And both of:
\itemize{
\item \code{variable} | \emph{character} | a string label grouping one concept code into a larger variable definition
\item \code{domain} | \emph{character} | the name of the CDM table where the concept can be found
}

For certain PCORnet applications, it should also contain
\itemize{
\item \code{vocabulary_id} | \emph{character} | the vocabulary of the code, which should match what is listed in the domain table's \code{vocabulary_field}
}

To see an example of this file structure, see \code{?sensitivityselectioncriteria::ssc_outcome_file}}

\item{domain_tbl}{\emph{tabular input} || defaults to the internal \code{ssc_domain_file}

A table or CSV file defining the domains to be assessed both for facts per patient year
computations & for any outcome variables. This input should contain:
\itemize{
\item \code{domain} | \emph{character} | a string label for the domain
\item \code{domain_tbl} | \emph{character} | the name of the CDM table where this domain is defined
\item \code{concept_field} | \emph{character}| the string name of the field in the domain table where the concepts are located
\item \code{date_field} | \emph{character} |  the name of the field in the domain table with the date that should be used for temporal filtering
\item \code{vocabulary_field} | \emph{character} | for PCORnet applications, the name of the field in the domain table with a vocabulary identifier to differentiate concepts from one another (ex: dx_type); can be set to NA for OMOP applications
\item \code{filter_logic} | \emph{character} | logic to be applied to the domain_tbl in order to achieve the definition of interest; should be written as if you were applying it in a \code{dplyr::filter} command in R
}

To see an example of this file structure, see \code{?sensitivityselectioncriteria::ssc_domain_file}}

\item{domain_select}{\emph{string or vector} || defaults to all internal \code{ssc_domain_file} domains

A string or vector of domain names, matching those listed in \code{domain_tbl},
that should be used in the computation of facts per patient year

This vector does NOT need to include domains used in the computation for
outcomes, as those will be accessed from the \code{domain_tbl} separately}
}
\value{
This function will return a dataframe summarizing fact and
demographic distributions for each of the provided cohort definitions.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
This is a plausibility module that will characterize user-provided base cohort
and alternate cohort definitions using a selection of continuous and categorical
variables. It will then use these values to compare each alternate definition to
the base cohort and evaluate how changes in the cohort definition impact the
patient population. Users have the option to provide definitions for several
variable types:
\itemize{
\item \code{demographic_mappings}: define a set of demographic variables of interest
\item \code{domain_tbl}: define domain definitions to assess patient fact density + utilization
(similar to Patient Facts module)
\item \code{specialty_concepts}: define a set of specialty concepts to evaluate specialty
care sought out by the cohort members
\item \code{outcome_concepts}: define study outcome variables
Sample versions of these input files are available with \verb{sensitivityselectioncriteria::}.
This module is compatible with both the OMOP and PCORnet CDMs based on the user's
selection.
}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'sensitivityselectioncriteria'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'ssc_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

## silence SQL trace for this example
config('db_trace', FALSE)

#' Build mock base study cohort
base_cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Build mock alternate study cohort
alt_cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  head(100) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Prepare input tables
ssc_domain_tbl <- dplyr::tibble(domain = c('all conditions', 'outpatient visits'),
                                domain_tbl = c('condition_occurrence', 'visit_occurrence'),
                                concept_field = c('condition_concept_id', 'visit_concept_id'),
                                date_field = c('condition_start_date', 'visit_start_date'),
                                vocabulary_field = c(NA, NA),
                                filter_logic = c(NA, 'visit_concept_id == 9202'))

ssc_outcome_tbl <- read_codeset('dx_hypertension') \%>\%
  dplyr::mutate(variable = 'hypertension', domain = 'all conditions')

#' Execute `ssc_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
ssc_process_example <- ssc_process(base_cohort = base_cohort,
                                   alt_cohorts = list('Sample Alternate' = alt_cohort),
                                   omop_or_pcornet = 'omop',
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   domain_tbl = ssc_domain_tbl,
                                   domain_select = c('all conditions', 'outpatient visits'),
                                   outcome_concepts = ssc_outcome_tbl) \%>\%
  suppressMessages()

ssc_process_example

#' Execute `ssc_output` function
ssc_output_example <- ssc_output(process_output = ssc_process_example,
                                 alt_cohort_filter = 'Sample Alternate') \%>\%
  suppressMessages()

ssc_output_example[[1]]
ssc_output_example[[2]]

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(ssc_output_example[[2]])
}
