% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_omop.R
\name{ssc_process_omop}
\alias{ssc_process_omop}
\title{Sensitivity to Selection Criteria -- OMOP}
\usage{
ssc_process_omop(
  base_cohort,
  alt_cohorts,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  person_tbl = cdm_tbl("person"),
  visit_tbl = cdm_tbl("visit_occurrence"),
  provider_tbl = cdm_tbl("provider"),
  care_site_tbl = cdm_tbl("care_site"),
  demographic_mappings = sensitivityselectioncriteria::ssc_omop_demographics,
  specialty_concepts = NULL,
  outcome_concepts = NULL,
  domain_tbl = sensitivityselectioncriteria::ssc_domain_file,
  domain_select = sensitivityselectioncriteria::ssc_domain_file \%>\% distinct(domain)
    \%>\% pull()
)
}
\arguments{
\item{base_cohort}{a dataframe including all patients who meet the base cohort definition
should have the columns site | person_id | start_date | end_date}

\item{alt_cohorts}{a list or named list of dataframes with patients meeting alternative
cohort definitions; if names are not provided, numbers will be assigned to each
cohort definition for labelling purposes}

\item{multi_or_single_site}{direction to determine what kind of check to run
string that is either \code{multi} or \code{single}}

\item{anomaly_or_exploratory}{direction to determine what kind of check to run; a string
that is either \code{anomaly} or \code{exploratory}}

\item{person_tbl}{CDM \code{person} table}

\item{visit_tbl}{CDM \code{visit_occurrence} table}

\item{provider_tbl}{the CDM table with provider & provider specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{care_site_tbl}{the CDM table with care site / facility & care site / facility specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{demographic_mappings}{table defining how demographic elements should be defined
if NULL, the default demographic mappings for the CDM will be used
(\code{ssc_omop_demographics})
otherwise, the user provided table will be used}

\item{specialty_concepts}{a concept set with provider specialty concepts of interest
to be used to identify specialty visits}

\item{outcome_concepts}{a concept set with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{variable} | \code{domain}

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    where domain matches a domain listed in the `domain_defs` table
}\if{html}{\out{</div>}}}

\item{domain_tbl}{a table with domain definitions with the following columns:
domain, domain_tbl, concept_field, date_field, filter_logic}

\item{domain_select}{a vector of domain names that should be used in the computation
of facts per domain per year of follow up

\if{html}{\out{<div class="sourceCode">}}\preformatted{                 this vector does NOT need to include domains used in the computation for
                 outcomes, as those will be accessed from the table directly
}\if{html}{\out{</div>}}}
}
\value{
a dataframe summarizing fact and demographic distributions for each of the provided
cohort definitions; for both anomaly detection outputs, the standardized mean difference
between the base cohort and each alternate cohort definition is computed for each
of the variables of interest
}
\description{
Sensitivity to Selection Criteria -- OMOP
}
\keyword{internal}
