% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_pcornet.R
\name{ssc_process_pcornet}
\alias{ssc_process_pcornet}
\title{Sensitivity to Selection Criteria -- PCORnet}
\usage{
ssc_process_pcornet(
  base_cohort,
  alt_cohorts,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  person_tbl = cdm_tbl("demographic"),
  visit_tbl = cdm_tbl("encounter"),
  provider_tbl = cdm_tbl("provider"),
  care_site_tbl = cdm_tbl("encounter"),
  black_codes = c("03"),
  white_codes = c("05"),
  asian_codes = c("02"),
  mixrace_codes = c("06"),
  unknown_codes = c("NI", "UN", "07"),
  other_codes = c("OT", "01", "04"),
  hispanic_codes = c("Y"),
  female_codes = c("F"),
  specialty_concepts = NULL,
  outcome_concepts = NULL,
  domain_tbl = sensitivityselectioncriteria::ssc_domain_file,
  domain_select = c("inpatient_visits", "outpatient_visits", "emergency_visits",
    "prescription_medications", "diagnoses")
)
}
\arguments{
\item{base_cohort}{a dataframe including all patients who meet the base cohort definition
should have the columns site | person_id | start_date | end_date}

\item{alt_cohorts}{a list or named list of dataframes with patients meeting alternative
cohort definitions; if names are not provided, numbers will be assigned to each
cohort definition for labelling purposes}

\item{multi_or_single_site}{direction to determine what kind of check to run
string that is either \code{multi} or \code{single}}

\item{anomaly_or_exploratory}{direction to determine what kind of check to run; a string
that is either \code{anomaly} or \code{exploratory}}

\item{person_tbl}{CDM \code{demographic} table}

\item{visit_tbl}{CDM \code{encounter} table}

\item{provider_tbl}{the CDM table with provider & provider specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{care_site_tbl}{the CDM table with care site / facility & care site / facility specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{black_codes}{list of codes that indicate that a patient is Black/African-American
defaults to standard PCORnet vocabulary -- \code{03}}

\item{white_codes}{list of codes that indicate that a patient is White/Caucasian
defaults to standard PCORnet vocabulary -- \code{05}}

\item{asian_codes}{list of codes that indicate that a patient is Asian
defaults to standard PCORnet vocabulary -- \code{02}}

\item{mixrace_codes}{list of codes that indicate that a patient is Mixed Race
defaults to standard PCORnet vocabulary -- \code{06}}

\item{unknown_codes}{list of codes that indicate that a patient's race is Unknown
defaults to standard PCORnet vocabulary -- \code{NI}, \code{UN}, \code{07}}

\item{other_codes}{list of codes that indicate that a patient's race is Unknown
defaults to -- \code{OT}, \code{01}, \code{04}}

\item{hispanic_codes}{list of codes that indicate that a patient is Hispanic or Latino
defaults to standard PCORnet vocabulary -- \code{Y}}

\item{female_codes}{list of codes that inidicate that a patient is Female
defaults to standard PCORnet vocabulary -- \code{F}}

\item{specialty_concepts}{a concept set with provider specialty concepts of interest
to be used to identify specialty visits}

\item{outcome_concepts}{a concept set with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{variable} | \code{domain}

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    where domain matches a domain listed in the `domain_defs` table
}\if{html}{\out{</div>}}}

\item{domain_tbl}{a table with domain definitions with the following columns:
domain, domain_tbl, concept_field, date_field, vocabulary_field, filter_logic}

\item{domain_select}{a vector of domain names that should be used in the computation
of facts per domain per year of follow up

\if{html}{\out{<div class="sourceCode">}}\preformatted{                 this vector does NOT need to include domains used in the computation for
                 outcomes, as those will be accessed from the table directly
}\if{html}{\out{</div>}}}
}
\value{
a dataframe summarizing fact and demographic distributions for each of the provided
cohort definitions; for both anomaly detection outputs, the standardized mean difference
between the base cohort and each alternate cohort definition is computed for each
of the variables of interest
}
\description{
Sensitivity to Selection Criteria -- PCORnet
}
