% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_output_subfuncs.R
\name{ssc_ss_anom_cs}
\alias{ssc_ss_anom_cs}
\title{\emph{Single Site, Anomaly Detection, Cross-Sectional}}
\usage{
ssc_ss_anom_cs(process_output)
}
\arguments{
\item{process_output}{the output from ssc_process}
}
\value{
a heat map displaying the standardized mean difference between
the base cohort and each of the alternate cohorts
}
\description{
\emph{Single Site, Anomaly Detection, Cross-Sectional}
}
