% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_output_subfuncs.R
\name{ssc_ss_exp_cs}
\alias{ssc_ss_exp_cs}
\title{\emph{Single Site, Exploratory, Cross-Sectional}}
\usage{
ssc_ss_exp_cs(summary_output, cohort_overlap, alt_cohort_filter)
}
\arguments{
\item{summary_output}{the summary dataframe output by ssc_process}

\item{cohort_overlap}{the cohort overlap dataframe output by ssc_process}

\item{alt_cohort_filter}{an vector indicating which alternate cohorts should be
displayed; only 3 are allowed at once to maintain visibility
in the graph}
}
\value{
an UpSet graph illustrating the overlap in patients between the two cohorts
a bar graph showing the difference between each alternate cohort and the base cohort
}
\description{
\emph{Single Site, Exploratory, Cross-Sectional}
}
