% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs_pcornet.R
\name{compare_cohort_def_pcnt}
\alias{compare_cohort_def_pcnt}
\title{Reformat cohort definition tables}
\usage{
compare_cohort_def_pcnt(
  base_cohort,
  alt_cohorts,
  multi_or_single_site,
  person_tbl = cdm_tbl("demographic"),
  visit_tbl = cdm_tbl("encounter"),
  provider_tbl = NULL,
  care_site_tbl = NULL,
  black_codes = c("03"),
  white_codes = c("05"),
  asian_codes = c("02"),
  mixrace_codes = c("06"),
  unknown_codes = c("NI", "UN", "07"),
  other_codes = c("OT", "01", "04"),
  hispanic_codes = c("Y"),
  female_codes = c("F"),
  specialty_concepts = NULL,
  outcome_concepts = NULL,
  domain_defs = sensitivityselectioncriteria::ssc_domain_file,
  domain_select = c("inpatient_visits", "outpatient_visits", "emergency_visits",
    "other_visits", "all_px", "prescription_medications", "all_conds")
)
}
\arguments{
\item{base_cohort}{table with the base cohort & inclusion criteria; must include
site, person_id, start_date, end_date}

\item{alt_cohorts}{list of each table with an alternative inclusion criteria definition
applied to the base_cohort}

\item{multi_or_single_site}{direction to determine what kind of check to run
string that is either \code{multi} or \code{single}}

\item{person_tbl}{the CDM table that contains patient information}

\item{visit_tbl}{the CDM table that contains visit/encounter information}

\item{provider_tbl}{the CDM table with provider & provider specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{care_site_tbl}{the CDM table with care site / facility & care site / facility specialty information
only used if \code{specialty_concepts} are provided
if provider_tbl & care_site_tbl are both not null, provider specialty is
prioritized}

\item{black_codes}{list of codes that indicate that a patient is Black/African-American
defaults to standard OMOP vocabulary -- \code{8516}}

\item{white_codes}{list of codes that indicate that a patient is White/Caucasian
defaults to standard OMOP vocabulary -- \code{8527}}

\item{asian_codes}{list of codes that indicate that a patient is Asian
defaults to standard OMOP vocabulary -- \code{8515}}

\item{mixrace_codes}{list of codes that indicate that a patient is Mixed Race
defaults to standard OMOP vocabulary -- \code{44814659}}

\item{unknown_codes}{list of codes that indicate that a patient's race is Unknown
defaults to standard OMOP vocabulary -- \code{44814660}, \code{44814650}, \code{44814653}}

\item{other_codes}{list of codes that indicate that a patient's race is Unknown
defaults to -- \code{44814649}, \code{8657}, \code{8557}}

\item{hispanic_codes}{list of codes that indicate that a patient is Hispanic or Latino
defaults to standard OMOP vocabulary -- \code{38003563}}

\item{female_codes}{list of codes that inidicate that a patient is Female
defaults to standard OMOP vocabulary -- \code{8532}}

\item{specialty_concepts}{a concept set with provider specialty concepts of interest
to be used to identify specialty visits}

\item{outcome_concepts}{a concept set with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{variable} | \code{domain}

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    where domain matches a domain listed in the `domain_defs` table
}\if{html}{\out{</div>}}}

\item{domain_defs}{a table with domain definitions with the following columns:
domain, domain_tbl, concept_field, date_field, filter_logic}

\item{domain_select}{a vector of domain names that should be used in the computation
of facts per domain per year of follow up

\if{html}{\out{<div class="sourceCode">}}\preformatted{                 this vector does NOT need to include domains used in the computation for
                 outcomes, as those will be accessed from the table directly
}\if{html}{\out{</div>}}}
}
\value{
one combined dataframe with one row per patient in the original cohort
with flags to show which cohort definitions apply to the patient.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    flag columns will include `base_cohort` and `alt_cohort_#` columns with numbers
    appended that correspond to the tables position in the `def_tbls` list

    `base_cohort` should equal 1 for all patients
}\if{html}{\out{</div>}}
}
\description{
Reformat cohort definition tables
}
