% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_process_subfuncs_omop.R
\name{compute_demographic_summary_omop}
\alias{compute_demographic_summary_omop}
\title{Compute patient level fact summary}
\usage{
compute_demographic_summary_omop(
  cohort_tbl,
  site_col,
  person_tbl = cdm_tbl("person"),
  visit_tbl = cdm_tbl("visit_occurrence"),
  demographic_mappings = sensitivityselectioncriteria::ssc_omop_demographics
)
}
\arguments{
\item{cohort_tbl}{table with the original cohort & inclustion criteria; must include
site, person_id, start_date, end_date}

\item{site_col}{the column in the cohort_tbl with the site name(s)}

\item{person_tbl}{CDM \code{person} table}

\item{visit_tbl}{CDM \code{visit_occurrence} table}

\item{demographic_mappings}{table defining how demographic elements should be defined; defaults
to \code{ssc_omop_demographics}; any additional demographic elements must be found in the person_tbl}
}
\value{
one dataframe with one row for each patient with columns
to show which facts apply to each patient:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    fu, age_cohort_entry, age_first_visit, and each of the demographic
    elements found in demographic_mappings
}\if{html}{\out{</div>}}
}
\description{
Compute patient level fact summary
}
\keyword{internal}
