% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc_output.R
\name{ssc_output}
\alias{ssc_output}
\title{Sensitivity to Selection Criteria Output}
\usage{
ssc_output(process_output, output_function, alt_cohort_filter)
}
\arguments{
\item{process_output}{the output from \code{ssc_process}}

\item{output_function}{the name of the output function to be executed; this is provided in the message
printed to the console after \code{scv_process} is finished running}

\item{alt_cohort_filter}{single site exploratory no time only -- a vector with the names of
alternate cohorts to display on the graph; this should be limited
to 3 or less to maintain good visibility on the graph}
}
\value{
a graph visualizing the differences between the alternate cohort definitions and
the base cohort; see documentation for each output function for specifics
}
\description{
Sensitivity to Selection Criteria Output
}
